+++ src/main/java/com/sotatek/cardano/common/entity/MaTxMint.java
@@ -1,7 +1,7 @@
 package com.sotatek.cardano.common.entity;
 
 import com.sotatek.cardano.common.validation.Int65Type;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -47,7 +47,7 @@ public class MaTxMint extends BaseEntity {
   @Column(name = "quantity", nullable = false, precision = 20)
   @Int65Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal quantity;
+  private BigInteger quantity;
 
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
